requirement.txt
asn1crypto==0.23.0
certifi==2016.2.28
cffi==1.11.1
cryptography==2.0.3
Django==1.11
djangorestframework==3.7.0
idna==2.6
psycopg2==2.7.3.1
pycparser==2.18
PyJWT==1.5.3
pytz==2017.2
six==1.11.0
